Проект разделён на модули
